changed how common package variables 6ba6182 and 6428dc9 are defined
The method previously implemented using `.onLoad()` and `assign()`
was not exporting to the namespace and broke references within
the functions. The values could still be accessed, but not referred
to cleanly/safely using `hRvstAPI::[variable_name]`. They now exist
as hidden variables loaded and exported from params.R, however, I
may implement setting these as options in the future to make make
the wrapper more flexible depending on its application.